requirements.txt
# Core dependencies for OneDay agent simulation tests
langwatch-scenario>=0.7.13
pytest>=8.4.2
python-dotenv>=1.1.1
litellm>=1.0.0
requests>=2.31.0

# Google API dependencies
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

